+++ lib/src/usecases/create_locate.dart
@@ -1,0 +1,14 @@
+import '../domain/domain.dart';
+import 'package:fpdart/fpdart.dart';
+
+/// Create a new locate.
+class CreateLocate {
+  /// Creates a [CreateLocate] use case.
+  const CreateLocate(this._repository);
+
+  final ILocateRepository _repository;
+
+  /// Executes the use case.
+  TaskEither<DomainFailure, Locate> call(Locate locate) =>
+      _repository.create(locate);
+}